services/ai-analyst/requirements.txt
fastapi
uvicorn
agno
groq
python-dotenv
requests